+++ lib/arachni/data/session.rb
@@ -1,0 +1,47 @@
+=begin
+    Copyright 2010-2014 Tasos Laskos <tasos.laskos@gmail.com>
+    All rights reserved.
+=end
+
+module Arachni
+class Data
+
+# Data for {Arachni::Session}.
+#
+# @author Tasos "Zapotek" Laskos <tasos.laskos@gmail.com>
+class Session
+
+    # @return   [Hash]
+    attr_reader :configuration
+
+    def initialize
+        @configuration = {}
+    end
+
+    def statistics
+        {}
+    end
+
+    def dump( directory )
+        FileUtils.mkdir_p( directory )
+
+        File.open( "#{directory}/configuration", 'w' ) do |f|
+            f.write Marshal.dump( @configuration )
+        end
+    end
+
+    def self.load( directory )
+        session = new
+        session.configuration.merge! Marshal.load( IO.read( "#{directory}/configuration" ) )
+        session
+    end
+
+    def clear
+        @configuration.clear
+    end
+
+end
+
+end
+end
+